Fix minor glitches
Change the pidfile default from leaving it to pidfile_open(3), which
means /var/run/powerd++.pid to /var/run/powerd.pid. This ensures
that powerd and powerd++ are not run at the same time (at least with
the defaults in place).
Before detaching from the terminal the run_daemon() function called
update_freq() for one last chance to fail before running in the
background. Due to the changes in the last commit update_freq() no
longer guarantees to set the core frequency every time it is called.
To still make sure that frequency setting works this call is replaced
with a loop reading and setting the current frequency for each core
group.