Move installer complexity around
So far the installer relied on the port passing in a sed command
to perform propper path substitutions like for %%PREFIX%%.
The logic for constructing the sed command moved into the install.sh
file. The Makefile now passes in the individual variables.
Additionally the Makefile now passes in the .CURDIR and .OBJDIR variables
so the installer can work when an OBJDIR is present.