Streamline the update_freq() function
Make the frequency variables in the function more representative by
naming them oldfreq, wantfreq and newfreq. The wantfreq and newfreq
are a split of the previous freq variable. The wantfreq variable
represents the calculated frequency, the newfreq variable is that value
adjusted to within the bounds of the upper and lower frequency boundaries.

The oldfreq and wantfreq value is reported in foreground mode.